requirements.txt
azure-cognitiveservices-speech==1.23.0
azure-common==1.1.28
azure-core==1.25.1
azure-mgmt-core==1.3.2
azure-mgmt-resource==21.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
